verilog: tools: ls: symbol-table-handler: Moved LSP URI to Path converters to separate lib
