+++ verilog/tools/ls/symbol-table-handler.h
@@ -30,15 +30,6 @@
 
 namespace verilog {
 
-// Converts file:// scheme entries to actual system paths.
-// If other scheme is provided, method returns empty string_view.
-// TODO (glatosinski) current resolving of LSP URIs is very naive
-// and supports only narrow use cases of file:// specifier.
-absl::string_view LSPUriToPath(absl::string_view uri);
-
-// Converts filesystem paths to file:// scheme entries.
-std::string PathToLSPUri(absl::string_view path);
-
 // Looks for FileList file for SymbolTableHandler
 std::string FindFileList(absl::string_view current_dir);
